move around app source
