+++ ooiui/core/app/science.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 '''
-ooiui.science.app
+ooiui.core.app.science
 
 Defines the application for the Science UI
 '''
@@ -12,7 +12,7 @@ app = Flask(__name__, static_url_path='')
 
 cache = Cache(app, config={'CACHE_TYPE': CACHE_TYPE})
 
-import ooiui.science.routes
+import ooiui.core.routes.science
 
 
 if __name__ == '__main__':
