+++ ooiui/templates/common/sample.html
@@ -1,0 +1,23 @@
+{% extends "base.html" %}
+
+{% block title %}
+  <title>Sample</title>
+{% endblock %}
+
+{% block head %}
+  
+  <!--
+  SOME CONTENT
+  -->
+
+
+{% endblock %}
+
+{% block body %}
+<div class="container">
+  <div class="row">
+    SAMPLE
+  </div>
+</div>
+{% endblock %}
+
